to typescript ok (remove all "menu navs" because it must be redone)
